initial commit by michaelgwelch
This initial commit copied the _data/contact.yml file
from the gem. I did this because I discovered from the
chirpy repo that they had support for mastodon, but I had
to copy it locally and uncomment that section. I then
also commented out email and twitter so they are not
displayed.